Reorganization of packages to enable better setup
- setup.py now includes slayerSNN which incorporates all files found in
src
