+++ cuda/shiftKernels.h
