+++ cuda/slayerKernels.cu
