+++ cuda/slayerLoihiKernels.cu
