+++ cuda/spikeKernels.h
