+++ cuda/spikeLoihiKernels.h
