+++ setup.py
@@ -7,12 +7,12 @@ setup(
         CUDAExtension(
             name='slayerCuda',
             sources=[
-                'src/cuda/slayerKernels.cu'
+                'cuda/slayerKernels.cu'
             ],
             depends=[
-                'src/cuda/spikeKernels.h',
-                'src/cuda/convKernels.h',
-                'src/cuda/shiftKernels.h'
+                'cuda/spikeKernels.h',
+                'cuda/convKernels.h',
+                'cuda/shiftKernels.h'
             ],
             extra_compile_args={
                 'cxx': ['-g'],
@@ -23,19 +23,16 @@ setup(
     cmdclass={'build_ext': BuildExtension}
 )
 
-from setuptools import setup
-from torch.utils.cpp_extension import BuildExtension, CUDAExtension
-
 setup(
     name='slayerLoihiCuda',
     ext_modules=[
         CUDAExtension(
             name='slayerLoihiCuda',
             sources=[
-                'src/cuda/slayerLoihiKernels.cu'
+                'cuda/slayerLoihiKernels.cu'
             ],
             depends=[
-                'src/cuda/spikeLoihiKernels.h'
+                'cuda/spikeLoihiKernels.h'
             ],
             extra_compile_args={
                 'cxx': ['-g'],
@@ -44,4 +41,9 @@ setup(
         )
     ],
     cmdclass={'build_ext': BuildExtension}
+)
+
+setup(
+    name='slayerSNN',
+    packages=["slayerSNN"],
 )
