+++ slayerSNN/slayerParams.py
